schemeshard: split schemeshard__operation_common.cpp, move code from .h
