+++ ydb/core/tx/schemeshard/schemeshard__operation_common.cpp
@@ -82,9 +82,393 @@ THolder<TEvHive::TEvCreateTablet> CreateEvCreateTablet(TPathElement::TPtr target
     return ev;
 }
 
-
-namespace
+// TCreateParts
+//
+TCreateParts::TCreateParts(const TOperationId& id)
+    : OperationId(id)
 {
+    IgnoreMessages(DebugHint(), {});
+}
+
+bool TCreateParts::HandleReply(TEvHive::TEvAdoptTabletReply::TPtr& ev, TOperationContext& context) {
+    LOG_INFO_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+                DebugHint() << " HandleReply TEvAdoptTablet"
+                << ", at tabletId: " << context.SS->SelfTabletId());
+    LOG_DEBUG_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+                DebugHint() << " HandleReply TEvAdoptTablet"
+                << ", message% " << DebugReply(ev));
+
+    NIceDb::TNiceDb db(context.GetDB());
+
+    const TString& explain = ev->Get()->Record.GetExplain();
+    TTabletId tabletId = TTabletId(ev->Get()->Record.GetTabletID()); // global id from hive
+    auto shardIdx = context.SS->MakeLocalId(TLocalShardIdx(ev->Get()->Record.GetOwnerIdx())); // global id from hive
+    TTabletId hive = TTabletId(ev->Get()->Record.GetOrigin());
+
+    Y_ABORT_UNLESS(ui64(context.SS->SelfTabletId()) == ev->Get()->Record.GetOwner());
+
+    NKikimrProto::EReplyStatus status = ev->Get()->Record.GetStatus();
+    Y_VERIFY_S(status ==  NKikimrProto::OK || status == NKikimrProto::ALREADY,
+                "Unexpected status " << NKikimrProto::EReplyStatus_Name(status)
+                                    << " in AdoptTabletReply for tabletId " << tabletId
+                                    << " with explain " << explain);
+
+    // Note that HIVE might send duplicate TTxAdoptTabletReply in case of restarts
+    // So we just ignore the event if we cannot find the Tx or if it is in a different
+    // state
+
+    Y_ABORT_UNLESS(context.SS->ShardInfos.contains(shardIdx));
+
+    if (!context.SS->AdoptedShards.contains(shardIdx)) {
+        LOG_INFO_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+                    DebugHint() << " HandleReply TEvAdoptTablet"
+                    << " Got TTxAdoptTabletReply for shard but it is not present in AdoptedShards"
+                    << ", shardIdx: " << shardIdx
+                    << ", tabletId:" << tabletId);
+        return false;
+    }
+
+    TShardInfo& shardInfo = context.SS->ShardInfos[shardIdx];
+    Y_ABORT_UNLESS(shardInfo.TabletID == InvalidTabletId || shardInfo.TabletID == tabletId);
+
+    Y_ABORT_UNLESS(tabletId != InvalidTabletId);
+    shardInfo.TabletID = tabletId;
+    context.SS->TabletIdToShardIdx[tabletId] = shardIdx;
+
+    TTxState* txState = context.SS->FindTx(OperationId);
+    Y_ABORT_UNLESS(txState);
+    Y_ABORT_UNLESS(txState->State == TTxState::CreateParts);
+
+    txState->ShardsInProgress.erase(shardIdx);
+
+    context.SS->AdoptedShards.erase(shardIdx);
+    context.SS->PersistDeleteAdopted(db, shardIdx);
+    context.SS->PersistShardMapping(db, shardIdx, tabletId, shardInfo.PathId, OperationId.GetTxId(), shardInfo.TabletType);
+
+    context.OnComplete.UnbindMsgFromPipe(OperationId, hive, shardIdx);
+    context.OnComplete.ActivateShardCreated(shardIdx, OperationId.GetTxId());
+
+    // If all datashards have been created
+    if (txState->ShardsInProgress.empty()) {
+        context.SS->ChangeTxState(db, OperationId, TTxState::ConfigureParts);
+        return true;
+    }
+
+    return false;
+}
+
+bool TCreateParts::HandleReply(TEvHive::TEvCreateTabletReply::TPtr& ev, TOperationContext& context) {
+    LOG_INFO_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+                DebugHint() << " HandleReply TEvCreateTabletReply"
+                << ", at tabletId: " << context.SS->SelfTabletId());
+    LOG_DEBUG_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+                DebugHint() << " HandleReply TEvCreateTabletReply"
+                << ", message: " << DebugReply(ev));
+
+    NIceDb::TNiceDb db(context.GetDB());
+
+    auto shardIdx = TShardIdx(ev->Get()->Record.GetOwner(),
+                                TLocalShardIdx(ev->Get()->Record.GetOwnerIdx()));
+
+    auto tabletId = TTabletId(ev->Get()->Record.GetTabletID()); // global id from hive
+    auto hive = TTabletId(ev->Get()->Record.GetOrigin());
+
+    NKikimrProto::EReplyStatus status = ev->Get()->Record.GetStatus();
+
+    Y_VERIFY_S(status ==  NKikimrProto::OK
+                    || status == NKikimrProto::ALREADY
+                    || status ==  NKikimrProto::INVALID_OWNER
+                    || status == NKikimrProto::BLOCKED,
+                "Unexpected status " << NKikimrProto::EReplyStatus_Name(status)
+                                    << " in CreateTabletReply shard idx " << shardIdx << " tabletId " << tabletId);
+
+    if (status ==  NKikimrProto::BLOCKED) {
+        Y_ABORT_UNLESS(!context.SS->IsDomainSchemeShard);
+
+        LOG_NOTICE_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+                        DebugHint() << " CreateRequest BLOCKED "
+                                    << " at Hive: " << hive
+                                    << " msg: " << DebugReply(ev));
+
+        // do not unsubscribe message
+        // context.OnComplete.UnbindMsgFromPipe(OperationId, hive, shardIdx);
+
+        // just stay calm and hung until tenant schemeshard is deleted
+        return false;
+    }
+
+    TTxState& txState = *context.SS->FindTx(OperationId);
+
+    TShardInfo& shardInfo = context.SS->ShardInfos.at(shardIdx);
+    Y_ABORT_UNLESS(shardInfo.TabletID == InvalidTabletId || shardInfo.TabletID == tabletId);
+
+    if (status ==  NKikimrProto::INVALID_OWNER) {
+        auto redirectTo = TTabletId(ev->Get()->Record.GetForwardRequest().GetHiveTabletId());
+        Y_ABORT_UNLESS(redirectTo);
+        Y_ABORT_UNLESS(tabletId);
+
+        context.OnComplete.UnbindMsgFromPipe(OperationId, hive, shardIdx);
+
+        auto path = context.SS->PathsById.at(txState.TargetPathId);
+        auto request = CreateEvCreateTablet(path, shardIdx, context);
+
+        LOG_DEBUG_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+                    DebugHint() << " CreateRequest"
+                                << " Redirect from Hive: " << hive
+                                << " to Hive: " << redirectTo
+                                << " msg:  " << request->Record.DebugString());
+
+        context.OnComplete.BindMsgToPipe(OperationId, redirectTo, shardIdx, request.Release());
+        return false;
+    }
+
+    if (shardInfo.TabletID == InvalidTabletId) {
+        switch (shardInfo.TabletType) {
+        case ETabletType::DataShard:
+            context.SS->TabletCounters->Simple()[COUNTER_TABLE_SHARD_ACTIVE_COUNT].Add(1);
+            break;
+        case ETabletType::Coordinator:
+            context.SS->TabletCounters->Simple()[COUNTER_SUB_DOMAIN_COORDINATOR_COUNT].Add(1);
+            break;
+        case ETabletType::Mediator:
+            context.SS->TabletCounters->Simple()[COUNTER_SUB_DOMAIN_MEDIATOR_COUNT].Add(1);
+            break;
+        case ETabletType::Hive:
+            context.SS->TabletCounters->Simple()[COUNTER_SUB_DOMAIN_HIVE_COUNT].Add(1);
+            break;
+        case ETabletType::SysViewProcessor:
+            context.SS->TabletCounters->Simple()[COUNTER_SYS_VIEW_PROCESSOR_COUNT].Add(1);
+            break;
+        case ETabletType::StatisticsAggregator:
+            context.SS->TabletCounters->Simple()[COUNTER_STATISTICS_AGGREGATOR_COUNT].Add(1);
+            break;
+        case ETabletType::BackupController:
+            context.SS->TabletCounters->Simple()[COUNTER_BACKUP_CONTROLLER_TABLET_COUNT].Add(1);
+            break;
+        default:
+            break;
+        }
+    }
+
+    Y_ABORT_UNLESS(tabletId != InvalidTabletId);
+    shardInfo.TabletID = tabletId;
+    context.SS->TabletIdToShardIdx[tabletId] = shardIdx;
+
+    Y_ABORT_UNLESS(OperationId.GetTxId() == shardInfo.CurrentTxId);
+
+    txState.ShardsInProgress.erase(shardIdx);
+
+    context.SS->PersistShardMapping(db, shardIdx, tabletId, shardInfo.PathId, OperationId.GetTxId(), shardInfo.TabletType);
+    context.OnComplete.UnbindMsgFromPipe(OperationId, hive, shardIdx);
+    context.OnComplete.ActivateShardCreated(shardIdx, OperationId.GetTxId());
+
+    // If all datashards have been created
+    if (txState.ShardsInProgress.empty()) {
+        context.SS->ChangeTxState(db, OperationId, TTxState::ConfigureParts);
+        return true;
+    }
+
+    return false;
+}
+
+THolder<TEvHive::TEvAdoptTablet> TCreateParts::AdoptRequest(TShardIdx shardIdx, TOperationContext& context) {
+    Y_ABORT_UNLESS(context.SS->AdoptedShards.contains(shardIdx));
+    auto& adoptedShard = context.SS->AdoptedShards[shardIdx];
+    auto& shard = context.SS->ShardInfos[shardIdx];
+
+    THolder<TEvHive::TEvAdoptTablet> ev = MakeHolder<TEvHive::TEvAdoptTablet>(
+        ui64(shard.TabletID),
+        adoptedShard.PrevOwner, ui64(adoptedShard.PrevShardIdx),
+        shard.TabletType,
+        ui64(context.SS->SelfTabletId()), ui64(shardIdx.GetLocalId()));
+
+    LOG_DEBUG_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+                DebugHint() << " AdoptRequest"
+                << " Event to Hive: " << ev->Record.DebugString().c_str());
+
+    return ev;
+}
+
+bool TCreateParts::ProgressState(TOperationContext& context) {
+    TTabletId ssId = context.SS->SelfTabletId();
+
+    TTxState* txState = context.SS->FindTx(OperationId);
+    Y_ABORT_UNLESS(txState);
+
+    LOG_INFO_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+                DebugHint() << " ProgressState"
+                            << ", operation type: " << TTxState::TypeName(txState->TxType)
+                            << ", at tablet" << ssId);
+
+    if (txState->TxType == TTxState::TxDropTable
+        || txState->TxType == TTxState::TxAlterTable
+        || txState->TxType == TTxState::TxBackup
+        || txState->TxType == TTxState::TxRestore) {
+        if (NTableState::CheckPartitioningChangedForTableModification(*txState, context)) {
+            LOG_INFO_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+                        DebugHint() << " ProgressState"
+                                    << " SourceTablePartitioningChangedForModification"
+                                    << ", tx type: " << TTxState::TypeName(txState->TxType));
+            NTableState::UpdatePartitioningForTableModification(OperationId, *txState, context);
+        }
+    } else if (txState->TxType == TTxState::TxCopyTable) {
+        if (NTableState::SourceTablePartitioningChangedForCopyTable(*txState, context)) {
+            LOG_INFO_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+                        DebugHint() << " ProgressState"
+                        << " SourceTablePartitioningChangedForCopyTable"
+                        << ", tx type: " << TTxState::TypeName(txState->TxType));
+            NTableState::UpdatePartitioningForCopyTable(OperationId, *txState, context);
+        }
+    }
+
+    txState->ClearShardsInProgress();
+
+    bool nothingToDo = true;
+    for (const auto& shard : txState->Shards) {
+        if (shard.Operation != TTxState::CreateParts) {
+            continue;
+        }
+        nothingToDo = false;
+
+        if (context.SS->AdoptedShards.contains(shard.Idx)) {
+            auto ev = AdoptRequest(shard.Idx, context);
+            context.OnComplete.BindMsgToPipe(OperationId, context.SS->GetGlobalHive(context.Ctx), shard.Idx, ev.Release());
+        } else {
+            auto path = context.SS->PathsById.at(txState->TargetPathId);
+            auto ev = CreateEvCreateTablet(path, shard.Idx, context);
+
+            auto hiveToRequest = context.SS->ResolveHive(shard.Idx, context.Ctx);
+
+            LOG_DEBUG_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+                        DebugHint() << " CreateRequest"
+                                    << " Event to Hive: " << hiveToRequest
+                                    << " msg:  "<< ev->Record.DebugString().c_str());
+
+            context.OnComplete.BindMsgToPipe(OperationId, hiveToRequest, shard.Idx, ev.Release());
+        }
+        context.OnComplete.RouteByShardIdx(OperationId, shard.Idx);
+    }
+
+    if (nothingToDo) {
+        LOG_DEBUG_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+                    DebugHint() << " ProgressState"
+                                << " no shards to create, do next state");
+
+        NIceDb::TNiceDb db(context.GetDB());
+        context.SS->ChangeTxState(db, OperationId, TTxState::ConfigureParts);
+        return true;
+    }
+
+    txState->UpdateShardsInProgress(TTxState::CreateParts);
+    return false;
+}
+
+// TDeleteParts
+//
+TDeleteParts::TDeleteParts(const TOperationId& id, TTxState::ETxState nextState)
+    : OperationId(id)
+    , NextState(nextState)
+{
+    IgnoreMessages(DebugHint(), {});
+}
+
+void TDeleteParts::DeleteShards(TOperationContext& context) {
+    const auto* txState = context.SS->FindTx(OperationId);
+
+    // Initiate asynchronous deletion of all shards
+    for (const auto& shard : txState->Shards) {
+        context.OnComplete.DeleteShard(shard.Idx);
+    }
+}
+
+bool TDeleteParts::ProgressState(TOperationContext& context) {
+    LOG_INFO_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+        "[" << context.SS->SelfTabletId() << "] " << DebugHint() << " ProgressState");
+    DeleteShards(context);
+
+    NIceDb::TNiceDb db(context.GetDB());
+    context.SS->ChangeTxState(db, OperationId, NextState);
+    return true;
+}
+
+// TDeletePartsAndDone
+//
+TDeletePartsAndDone::TDeletePartsAndDone(const TOperationId& id)
+    : TDeleteParts(id)
+{
+    Y_UNUSED(NextState);
+}
+
+bool TDeletePartsAndDone::ProgressState(TOperationContext& context) {
+    LOG_INFO_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+        "[" << context.SS->SelfTabletId() << "] " << DebugHint() << " ProgressState");
+    DeleteShards(context);
+
+    context.OnComplete.DoneOperation(OperationId);
+    return true;
+}
+
+// TDone
+//
+TDone::TDone(const TOperationId& id)
+    : OperationId(id)
+{
+    IgnoreMessages(DebugHint(), AllIncomingEvents());
+}
+
+bool TDone::ProgressState(TOperationContext& context) {
+    LOG_INFO_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+        "[" << context.SS->SelfTabletId() << "] " << DebugHint() << " ProgressState");
+
+    const auto* txState = context.SS->FindTx(OperationId);
+
+    const auto& pathId = txState->TargetPathId;
+    Y_ABORT_UNLESS(context.SS->PathsById.contains(pathId));
+    TPathElement::TPtr path = context.SS->PathsById.at(pathId);
+    Y_VERIFY_S(path->PathState != TPathElement::EPathState::EPathStateNoChanges, "with context"
+        << ", PathState: " << NKikimrSchemeOp::EPathState_Name(path->PathState)
+        << ", PathId: " << path->PathId
+        << ", PathName: " << path->Name);
+
+    if (path->IsPQGroup() && txState->IsCreate()) {
+        TPathElement::TPtr parentDir = context.SS->PathsById.at(path->ParentPathId);
+        // can't be removed until KIKIMR-8861
+        // at least lets wrap it into condition
+        // it helps to actualize PATHSTATE inside children listing
+        context.SS->ClearDescribePathCaches(parentDir);
+    }
+
+    if (txState->IsDrop()) {
+        context.OnComplete.ReleasePathState(OperationId, path->PathId, TPathElement::EPathState::EPathStateNotExist);
+    } else {
+        context.OnComplete.ReleasePathState(OperationId, path->PathId, TPathElement::EPathState::EPathStateNoChanges);
+    }
+
+    if (txState->SourcePathId != InvalidPathId) {
+        Y_ABORT_UNLESS(context.SS->PathsById.contains(txState->SourcePathId));
+        TPathElement::TPtr srcPath = context.SS->PathsById.at(txState->SourcePathId);
+        if (srcPath->PathState == TPathElement::EPathState::EPathStateCopying) {
+            context.OnComplete.ReleasePathState(OperationId, srcPath->PathId, TPathElement::EPathState::EPathStateNoChanges);
+        }
+    }
+
+    // OlapStore tracks all tables that are under operation, make sure to unlink
+    if (context.SS->ColumnTables.contains(pathId)) {
+        auto tableInfo = context.SS->ColumnTables.at(pathId);
+        if (!tableInfo->IsStandalone()) {
+            const auto storePathId = tableInfo->GetOlapStorePathIdVerified();
+            if (context.SS->OlapStores.contains(storePathId)) {
+                auto storeInfo = context.SS->OlapStores.at(storePathId);
+                storeInfo->ColumnTablesUnderOperation.erase(pathId);
+            }
+        }
+    }
+
+    context.OnComplete.DoneOperation(OperationId);
+    return true;
+}
+
+namespace {
 
 template <typename T, typename TFuncCheck, typename TFuncToString>
 bool CollectProposeTxResults(
@@ -158,7 +542,9 @@ bool CollectProposeTxResults(
 
 } // anonymous namespace
 
-bool NTableState::CollectProposeTransactionResults(
+namespace NTableState {
+
+bool CollectProposeTransactionResults(
         const NKikimr::NSchemeShard::TOperationId &operationId,
         const TEvDataShard::TEvProposeTransactionResult::TPtr &ev,
         NKikimr::NSchemeShard::TOperationContext &context)
@@ -174,7 +560,7 @@ bool NTableState::CollectProposeTransactionResults(
     return CollectProposeTxResults(ev, operationId, context, prepared, toString);
 }
 
-bool NTableState::CollectProposeTransactionResults(
+bool CollectProposeTransactionResults(
         const NKikimr::NSchemeShard::TOperationId& operationId,
         const TEvColumnShard::TEvProposeTransactionResult::TPtr& ev,
         NKikimr::NSchemeShard::TOperationContext& context)
@@ -190,7 +576,7 @@ bool NTableState::CollectProposeTransactionResults(
     return CollectProposeTxResults(ev, operationId, context, prepared, toString);
 }
 
-bool NTableState::CollectSchemaChanged(
+bool CollectSchemaChanged(
         const TOperationId& operationId,
         const TEvDataShard::TEvSchemaChanged::TPtr& ev,
         TOperationContext& context)
@@ -263,7 +649,7 @@ bool NTableState::CollectSchemaChanged(
     return false;
 }
 
-void NTableState::AckAllSchemaChanges(const TOperationId &operationId, TTxState &txState, TOperationContext &context) {
+void AckAllSchemaChanges(const TOperationId &operationId, TTxState &txState, TOperationContext &context) {
     TTabletId ssId = context.SS->SelfTabletId();
 
     LOG_INFO_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
@@ -290,7 +676,7 @@ void NTableState::AckAllSchemaChanges(const TOperationId &operationId, TTxState 
     }
 }
 
-bool NTableState::CheckPartitioningChangedForTableModification(TTxState &txState, TOperationContext &context) {
+bool CheckPartitioningChangedForTableModification(TTxState &txState, TOperationContext &context) {
     Y_ABORT_UNLESS(context.SS->Tables.contains(txState.TargetPathId));
     TTableInfo::TPtr table = context.SS->Tables.at(txState.TargetPathId);
 
@@ -309,7 +695,7 @@ bool NTableState::CheckPartitioningChangedForTableModification(TTxState &txState
     return !shardIdxsLeft.empty();
 }
 
-void NTableState::UpdatePartitioningForTableModification(TOperationId operationId, TTxState &txState, TOperationContext &context) {
+void UpdatePartitioningForTableModification(TOperationId operationId, TTxState &txState, TOperationContext &context) {
     Y_ABORT_UNLESS(!txState.TxShardsListFinalized, "Rebuilding the list of shards must not happen twice");
 
     NIceDb::TNiceDb db(context.GetDB());
@@ -428,7 +814,7 @@ void NTableState::UpdatePartitioningForTableModification(TOperationId operationI
     txState.TxShardsListFinalized = true;
 }
 
-bool NTableState::SourceTablePartitioningChangedForCopyTable(const TTxState &txState, TOperationContext &context) {
+bool SourceTablePartitioningChangedForCopyTable(const TTxState &txState, TOperationContext &context) {
     Y_ABORT_UNLESS(txState.SourcePathId != InvalidPathId);
     Y_ABORT_UNLESS(txState.TargetPathId != InvalidPathId);
     const TTableInfo::TPtr srcTableInfo = *context.SS->Tables.FindPtr(txState.SourcePathId);
@@ -453,7 +839,7 @@ bool NTableState::SourceTablePartitioningChangedForCopyTable(const TTxState &txS
     return !srcShardIdxsLeft.empty();
 }
 
-void NTableState::UpdatePartitioningForCopyTable(TOperationId operationId, TTxState &txState, TOperationContext &context) {
+void UpdatePartitioningForCopyTable(TOperationId operationId, TTxState &txState, TOperationContext &context) {
     Y_ABORT_UNLESS(!txState.TxShardsListFinalized, "CopyTable can adjust partitioning only once");
 
     // Source table must not be altered or drop while we are performing copying. So we put it into a special state.
@@ -555,7 +941,7 @@ void NTableState::UpdatePartitioningForCopyTable(TOperationId operationId, TTxSt
     txState.TxShardsListFinalized = true;
 }
 
-TVector<TTableShardInfo> NTableState::ApplyPartitioningCopyTable(const TShardInfo &templateDatashardInfo, TTableInfo::TPtr srcTableInfo, TTxState &txState, TSchemeShard *ss) {
+TVector<TTableShardInfo> ApplyPartitioningCopyTable(const TShardInfo &templateDatashardInfo, TTableInfo::TPtr srcTableInfo, TTxState &txState, TSchemeShard *ss) {
     TVector<TTableShardInfo> dstPartitions = srcTableInfo->GetPartitions();
 
     // Source table must not be altered or drop while we are performing copying. So we put it into a special state.
@@ -580,6 +966,91 @@ TVector<TTableShardInfo> NTableState::ApplyPartitioningCopyTable(const TShardInf
     return dstPartitions;
 }
 
+// NTableState::TProposedWaitParts
+//
+TProposedWaitParts::TProposedWaitParts(TOperationId id, TTxState::ETxState nextState)
+    : OperationId(id)
+    , NextState(nextState)
+{
+    IgnoreMessages(DebugHint(),
+        { TEvHive::TEvCreateTabletReply::EventType
+        , TEvDataShard::TEvProposeTransactionResult::EventType
+        , TEvPrivate::TEvOperationPlan::EventType }
+    );
+}
+
+bool TProposedWaitParts::HandleReply(TEvDataShard::TEvSchemaChanged::TPtr& ev, TOperationContext& context) {
+    TTabletId ssId = context.SS->SelfTabletId();
+    const auto& evRecord = ev->Get()->Record;
+
+    LOG_INFO_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+                DebugHint() << " HandleReply TEvSchemaChanged"
+                            << " at tablet: " << ssId);
+    LOG_DEBUG_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+                DebugHint() << " HandleReply TEvSchemaChanged"
+                            << " at tablet: " << ssId
+                            << " message: " << evRecord.ShortDebugString());
+
+    if (!CollectSchemaChanged(OperationId, ev, context)) {
+        LOG_DEBUG_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+                    DebugHint() << " HandleReply TEvSchemaChanged"
+                                << " CollectSchemaChanged: false");
+        return false;
+    }
+
+    Y_ABORT_UNLESS(context.SS->FindTx(OperationId));
+    TTxState& txState = *context.SS->FindTx(OperationId);
+
+    if (!txState.ReadyForNotifications) {
+        LOG_DEBUG_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+                    DebugHint() << " HandleReply TEvSchemaChanged"
+                                << " ReadyForNotifications: false");
+        return false;
+    }
+
+    return true;
+}
+
+bool TProposedWaitParts::ProgressState(TOperationContext& context) {
+    TTabletId ssId = context.SS->SelfTabletId();
+
+    LOG_INFO_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+                    DebugHint() << " ProgressState"
+                    << " at tablet: " << ssId);
+
+    TTxState* txState = context.SS->FindTx(OperationId);
+
+    NIceDb::TNiceDb db(context.GetDB());
+
+    txState->ClearShardsInProgress();
+    for (TTxState::TShardOperation& shard : txState->Shards) {
+        if (shard.Operation < TTxState::ProposedWaitParts) {
+            shard.Operation = TTxState::ProposedWaitParts;
+            context.SS->PersistUpdateTxShard(db, OperationId, shard.Idx, shard.Operation);
+        }
+        Y_ABORT_UNLESS(context.SS->ShardInfos.contains(shard.Idx));
+        context.OnComplete.RouteByTablet(OperationId,  context.SS->ShardInfos.at(shard.Idx).TabletID);
+    }
+    txState->UpdateShardsInProgress(TTxState::ProposedWaitParts);
+
+    // Move all notifications that were already received
+    // NOTE: SchemeChangeNotification is sent form DS after it has got PlanStep from coordinator and the schema tx has completed
+    // At that moment the SS might not have received PlanStep from coordinator yet (this message might be still on its way to SS)
+    // So we are going to accumulate SchemeChangeNotification that are received before this Tx switches to WaitParts state
+    txState->AcceptPendingSchemeNotification();
+
+    // Got notifications from all datashards?
+    if (txState->ShardsInProgress.empty()) {
+        AckAllSchemaChanges(OperationId, *txState, context);
+        context.SS->ChangeTxState(db, OperationId, NextState);
+        return true;
+    }
+
+    return false;
+}
+
+}  // namespace NTableState
+
 namespace NPQState {
 
 bool CollectProposeTransactionResults(const TOperationId& operationId,
@@ -597,319 +1068,7 @@ bool CollectProposeTransactionResults(const TOperationId& operationId,
     return CollectProposeTxResults(ev, operationId, context, prepared, toString);
 }
 
-bool CollectPQConfigChanged(const TOperationId& operationId,
-                            const TEvPersQueue::TEvProposeTransactionResult::TPtr& ev,
-                            TOperationContext& context)
-{
-    Y_ABORT_UNLESS(context.SS->FindTx(operationId));
-    TTxState& txState = *context.SS->FindTx(operationId);
-
-    const auto& evRecord = ev->Get()->Record;
-    if (evRecord.GetStatus() == NKikimrPQ::TEvProposeTransactionResult::COMPLETE) {
-        const auto ssId = context.SS->SelfTabletId();
-        const TTabletId shardId(evRecord.GetOrigin());
-
-        const auto shardIdx = context.SS->MustGetShardIdx(shardId);
-        Y_ABORT_UNLESS(context.SS->ShardInfos.contains(shardIdx));
-
-        Y_ABORT_UNLESS(txState.State == TTxState::Propose);
-
-        txState.ShardsInProgress.erase(shardIdx);
-
-        LOG_DEBUG_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
-                    "CollectPQConfigChanged accept TEvPersQueue::TEvProposeTransactionResult"
-                    << ", operationId: " << operationId
-                    << ", shardIdx: " << shardIdx
-                    << ", shard: " << shardId
-                    << ", left await: " << txState.ShardsInProgress.size()
-                    << ", txState.State: " << TTxState::StateName(txState.State)
-                    << ", txState.ReadyForNotifications: " << txState.ReadyForNotifications
-                    << ", at schemeshard: " << ssId);
-    }
-
-    return txState.ShardsInProgress.empty();
-}
-
-bool CollectPQConfigChanged(const TOperationId& operationId,
-                            const TEvPersQueue::TEvProposeTransactionAttachResult::TPtr& ev,
-                            TOperationContext& context)
-{
-    Y_ABORT_UNLESS(context.SS->FindTx(operationId));
-    TTxState& txState = *context.SS->FindTx(operationId);
-
-    //
-    // The PQ tablet can perform a transaction and send a TEvProposeTransactionResult(COMPLETE) response.
-    // The SchemeShard tablet can restart at this point. After restarting at the TPropose step, it will
-    // send the TEvProposeTransactionAttach message to the PQ tablets. If the NODATA status is specified in
-    // the response TEvProposeTransactionAttachResult, then the PQ tablet has already completed the transaction.
-    // Otherwise, she continues to execute the transaction
-    //
-
-    const auto& evRecord = ev->Get()->Record;
-    if (evRecord.GetStatus() != NKikimrProto::NODATA) {
-        //
-        // If the PQ tablet returned something other than NODATA, then it continues to execute the transaction
-        //
-        return txState.ShardsInProgress.empty();
-    }
-
-    //
-    // Otherwise, she has already completed the transaction and has forgotten about it. Then we can
-    // remove PQ tablet from the list of shards
-    //
-
-    const auto ssId = context.SS->SelfTabletId();
-    const TTabletId shardId(evRecord.GetTabletId());
-
-    const auto shardIdx = context.SS->MustGetShardIdx(shardId);
-    Y_ABORT_UNLESS(context.SS->ShardInfos.contains(shardIdx));
-
-    Y_ABORT_UNLESS(txState.State == TTxState::Propose);
-
-    txState.ShardsInProgress.erase(shardIdx);
-
-    LOG_DEBUG_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
-                "CollectPQConfigChanged accept TEvPersQueue::TEvProposeTransactionAttachResult"
-                << ", operationId: " << operationId
-                << ", shardIdx: " << shardIdx
-                << ", shard: " << shardId
-                << ", left await: " << txState.ShardsInProgress.size()
-                << ", txState.State: " << TTxState::StateName(txState.State)
-                << ", txState.ReadyForNotifications: " << txState.ReadyForNotifications
-                << ", at schemeshard: " << ssId);
-
-    return txState.ShardsInProgress.empty();
-}
-
-bool TConfigureParts::HandleReply(TEvPersQueue::TEvProposeTransactionResult::TPtr& ev, TOperationContext& context)
-{
-    const TTabletId ssId = context.SS->SelfTabletId();
-
-    LOG_INFO_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
-               DebugHint() << " HandleReply TEvProposeTransactionResult"
-               << ", at schemeshard: " << ssId);
-
-    return CollectProposeTransactionResults(OperationId, ev, context);
-}
-
-THolder<TEvPersQueue::TEvProposeTransaction> TConfigureParts::MakeEvProposeTransaction(TTxId txId,
-                                                                                       const TTopicInfo& pqGroup,
-                                                                                       const TTopicTabletInfo& pqShard,
-                                                                                       const TString& topicName,
-                                                                                       const TString& topicPath,
-                                                                                       const std::optional<TBootstrapConfigWrapper>& bootstrapConfig,
-                                                                                       const TString& cloudId,
-                                                                                       const TString& folderId,
-                                                                                       const TString& databaseId,
-                                                                                       const TString& databasePath,
-                                                                                       TTxState::ETxType txType,
-                                                                                       const TOperationContext& context)
-{
-    auto event = MakeHolder<TEvPersQueue::TEvProposeTransactionBuilder>();
-    event->Record.SetTxId(ui64(txId));
-    ActorIdToProto(context.SS->SelfId(), event->Record.MutableSourceActor());
-
-    MakePQTabletConfig(context,
-                      *event->Record.MutableConfig()->MutableTabletConfig(),
-                       pqGroup,
-                       pqShard,
-                       topicName,
-                       topicPath,
-                       cloudId,
-                       folderId,
-                       databaseId,
-                       databasePath);
-    if (bootstrapConfig) {
-        Y_ABORT_UNLESS(txType == TTxState::TxCreatePQGroup);
-        event->PreSerializedData += bootstrapConfig->GetPreSerializedProposeTransaction();
-    }
-
-    LOG_DEBUG_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
-                "Propose configure PersQueue" <<
-                ", message: " << event->Record.ShortUtf8DebugString());
-
-    return event;
-}
-
-THolder<TEvPersQueue::TEvUpdateConfig> TConfigureParts::MakeEvUpdateConfig(TTxId txId,
-                                                                           const TTopicInfo& pqGroup,
-                                                                           const TTopicTabletInfo& pqShard,
-                                                                           const TString& topicName,
-                                                                           const TString& topicPath,
-                                                                           const std::optional<TBootstrapConfigWrapper>& bootstrapConfig,
-                                                                           const TString& cloudId,
-                                                                           const TString& folderId,
-                                                                           const TString& databaseId,
-                                                                           const TString& databasePath,
-                                                                           TTxState::ETxType txType,
-                                                                           const TOperationContext& context)
-{
-    auto event = MakeHolder<TEvPersQueue::TEvUpdateConfigBuilder>();
-    event->Record.SetTxId(ui64(txId));
-
-    MakePQTabletConfig(context,
-                       *event->Record.MutableTabletConfig(),
-                       pqGroup,
-                       pqShard,
-                       topicName,
-                       topicPath,
-                       cloudId,
-                       folderId,
-                       databaseId,
-                       databasePath);
-    if (bootstrapConfig) {
-        Y_ABORT_UNLESS(txType == TTxState::TxCreatePQGroup);
-        event->PreSerializedData += bootstrapConfig->GetPreSerializedUpdateConfig();
-    }
-
-    LOG_DEBUG_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
-                "Propose configure PersQueue" <<
-                ", message: " << event->Record.ShortUtf8DebugString());
-
-    return event;
-}
-
-bool TPropose::HandleReply(TEvPersQueue::TEvProposeTransactionResult::TPtr& ev, TOperationContext& context)
-{
-    const TTabletId ssId = context.SS->SelfTabletId();
-    const auto& evRecord = ev->Get()->Record;
-
-    LOG_INFO_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
-               DebugHint() << " HandleReply TEvProposeTransactionResult"
-               << " triggers early"
-               << ", at schemeshard: " << ssId
-               << " message# " << evRecord.ShortDebugString());
-
-    const bool collected = CollectPQConfigChanged(OperationId, ev, context);
-    LOG_DEBUG_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
-                DebugHint() << " HandleReply TEvProposeTransactionResult"
-                << " CollectPQConfigChanged: " << (collected ? "true" : "false"));
-
-    return TryPersistState(context);
-}
-
-bool TPropose::HandleReply(TEvPersQueue::TEvProposeTransactionAttachResult::TPtr& ev, TOperationContext& context)
-{
-    const auto ssId = context.SS->SelfTabletId();
-    const auto& evRecord = ev->Get()->Record;
-
-    LOG_INFO_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
-               DebugHint() << " HandleReply TEvProposeTransactionAttachResult"
-               << " triggers early"
-               << ", at schemeshard: " << ssId
-               << " message# " << evRecord.ShortDebugString());
-
-    const bool collected = CollectPQConfigChanged(OperationId, ev, context);
-    LOG_DEBUG_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
-                DebugHint() << " HandleReply TEvProposeTransactionAttachResult"
-                << " CollectPQConfigChanged: " << (collected ? "true" : "false"));
-
-    return TryPersistState(context);
-}
-
-void TPropose::PrepareShards(TTxState& txState, TSet<TTabletId>& shardSet, TOperationContext& context)
-{
-    txState.UpdateShardsInProgress();
-
-    for (const auto& shard : txState.Shards) {
-        const TShardIdx idx = shard.Idx;
-        //
-        // The operation involves the ETabletType::PersQueue and Tabletype::PersQueueReadBalancer shards.
-        // The program receives responses from PersQueueReadBalancer at the previous stage. At this stage,
-        // it only expects TEvProposeTransactionResult from PersQueue
-        //
-        if (shard.TabletType == ETabletType::PersQueue) {
-            const TTabletId tablet = context.SS->ShardInfos.at(idx).TabletID;
-
-            shardSet.insert(tablet);
-
-            //
-            // By this point, the SchemeShard tablet could restart and the actor ID changed. Therefore, we send
-            // the TEvProposeTransactionAttach message to the PQ tablets so that they recognize the new recipient
-            //
-            SendEvProposeTransactionAttach(idx, tablet, context);
-        } else {
-            txState.ShardsInProgress.erase(idx);
-        }
-    }
-}
-
-void TPropose::SendEvProposeTransactionAttach(TShardIdx shard, TTabletId tablet,
-                                              TOperationContext& context)
-{
-    auto event =
-        MakeHolder<TEvPersQueue::TEvProposeTransactionAttach>(ui64(tablet),
-                                                              ui64(OperationId.GetTxId()));
-    context.OnComplete.BindMsgToPipe(OperationId, tablet, shard, event.Release());
-}
-
-bool TPropose::CanPersistState(const TTxState& txState,
-                               TOperationContext& context)
-{
-    if (!txState.ShardsInProgress.empty()) {
-        LOG_DEBUG_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
-                    DebugHint() << " can't persist state: " <<
-                    "ShardsInProgress is not empty, remain: " << txState.ShardsInProgress.size());
-        return false;
-    }
-
-    PathId = txState.TargetPathId;
-    Path = context.SS->PathsById.at(PathId);
-
-    if (Path->StepCreated == InvalidStepId) {
-        LOG_DEBUG_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
-                    DebugHint() << " can't persist state: " <<
-                    "StepCreated is invalid");
-        return false;
-    }
-
-    return true;
-}
-
-void TPropose::PersistState(const TTxState& txState,
-                            TOperationContext& context) const
-{
-    NIceDb::TNiceDb db(context.GetDB());
-
-    if (txState.TxType == TTxState::TxCreatePQGroup) {
-        auto parentDir = context.SS->PathsById.at(Path->ParentPathId);
-        ++parentDir->DirAlterVersion;
-        context.SS->PersistPathDirAlterVersion(db, parentDir);
-        context.SS->ClearDescribePathCaches(parentDir);
-        context.OnComplete.PublishToSchemeBoard(OperationId, parentDir->PathId);
-    }
-
-    context.SS->ClearDescribePathCaches(Path);
-    context.OnComplete.PublishToSchemeBoard(OperationId, PathId);
-
-    TTopicInfo::TPtr pqGroup = context.SS->Topics[PathId];
-
-    NKikimrPQ::TPQTabletConfig tabletConfig = pqGroup->GetTabletConfig();
-    NKikimrPQ::TPQTabletConfig newTabletConfig = pqGroup->AlterData->GetTabletConfig();
-
-    pqGroup->FinishAlter();
-
-    context.SS->PersistPersQueueGroup(db, PathId, pqGroup);
-    context.SS->PersistRemovePersQueueGroupAlter(db, PathId);
-
-    context.SS->ChangeTxState(db, OperationId, TTxState::Done);
-}
-
-bool TPropose::TryPersistState(TOperationContext& context)
-{
-    TTxState* txState = context.SS->FindTx(OperationId);
-    Y_ABORT_UNLESS(txState);
-
-    if (!CanPersistState(*txState, context)) {
-        return false;
-    }
-
-    PersistState(*txState, context);
-
-    return true;
-}
-
-}
+}  // namespace NPQState
 
 TSet<ui32> AllIncomingEvents() {
     TSet<ui32> result;
@@ -923,7 +1082,9 @@ TSet<ui32> AllIncomingEvents() {
     return result;
 }
 
-void NForceDrop::CollectShards(const THashSet<TPathId>& paths, TOperationId operationId, TTxState *txState, TOperationContext &context) {
+namespace NForceDrop {
+
+void CollectShards(const THashSet<TPathId>& paths, TOperationId operationId, TTxState *txState, TOperationContext &context) {
     NIceDb::TNiceDb db(context.GetDB());
 
     auto shards = context.SS->CollectAllShards(paths);
@@ -950,7 +1111,7 @@ void NForceDrop::CollectShards(const THashSet<TPathId>& paths, TOperationId oper
     context.SS->PersistTxState(db, operationId);
 }
 
-void NForceDrop::ValidateNoTransactionOnPaths(TOperationId operationId, const THashSet<TPathId>& paths, TOperationContext &context) {
+void ValidateNoTransactionOnPaths(TOperationId operationId, const THashSet<TPathId>& paths, TOperationContext &context) {
     // No transaction should materialize in a subdomain that is being deleted --
     // -- all operations should be checking parent dir status at Propose stage.
     // However, it is better to verify that, just in case.
@@ -962,6 +1123,8 @@ void NForceDrop::ValidateNoTransactionOnPaths(TOperationId operationId, const TH
         Y_VERIFY_S(false, "unexpected transaction: " << otherTxId << " found on the subdomain being deleted by transaction " << operationId.GetTxId());
     }
 }
+
+}  // namespace NForceDrop
 
 void IncParentDirAlterVersionWithRepublishSafeWithUndo(const TOperationId& opId, const TPath& path, TSchemeShard* ss, TSideEffects& onComplete) {
     auto parent = path.Parent();
