+++ ydb/core/tx/schemeshard/schemeshard__operation_common_cdc_stream.cpp
@@ -1,0 +1,179 @@
+#include "schemeshard__operation_common.h"
+
+#include "schemeshard_private.h"
+#include <ydb/core/base/hive.h>
+#include <ydb/core/tx/datashard/datashard.h>
+
+
+namespace NKikimr::NSchemeShard::NCdcStreamState {
+
+namespace {
+
+bool IsExpectedTxType(TTxState::ETxType txType) {
+    switch (txType) {
+    case TTxState::TxCreateCdcStreamAtTable:
+    case TTxState::TxCreateCdcStreamAtTableWithInitialScan:
+    case TTxState::TxAlterCdcStreamAtTable:
+    case TTxState::TxAlterCdcStreamAtTableDropSnapshot:
+    case TTxState::TxDropCdcStreamAtTable:
+    case TTxState::TxDropCdcStreamAtTableDropSnapshot:
+        return true;
+    default:
+        return false;
+    }
+}
+
+}  // namespace anonymous
+
+
+// NCdcStreamState::TConfigurePartsAtTable
+//
+TConfigurePartsAtTable::TConfigurePartsAtTable(TOperationId id)
+    : OperationId(id)
+{
+    IgnoreMessages(DebugHint(), {});
+}
+
+bool TConfigurePartsAtTable::ProgressState(TOperationContext& context) {
+    LOG_INFO_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+                DebugHint() << " ProgressState"
+                            << ", at schemeshard: " << context.SS->SelfTabletId());
+
+    auto* txState = context.SS->FindTx(OperationId);
+    Y_ABORT_UNLESS(txState);
+    Y_ABORT_UNLESS(IsExpectedTxType(txState->TxType));
+    const auto& pathId = txState->TargetPathId;
+
+    if (NTableState::CheckPartitioningChangedForTableModification(*txState, context)) {
+        NTableState::UpdatePartitioningForTableModification(OperationId, *txState, context);
+    }
+
+    NKikimrTxDataShard::TFlatSchemeTransaction tx;
+    context.SS->FillSeqNo(tx, context.SS->StartRound(*txState));
+    FillNotice(pathId, tx, context);
+
+    txState->ClearShardsInProgress();
+    Y_ABORT_UNLESS(txState->Shards.size());
+
+    for (ui32 i = 0; i < txState->Shards.size(); ++i) {
+        const auto& idx = txState->Shards[i].Idx;
+        const auto datashardId = context.SS->ShardInfos[idx].TabletID;
+        auto ev = context.SS->MakeDataShardProposal(pathId, OperationId, tx.SerializeAsString(), context.Ctx);
+        context.OnComplete.BindMsgToPipe(OperationId, datashardId, idx, ev.Release());
+    }
+
+    txState->UpdateShardsInProgress(TTxState::ConfigureParts);
+    return false;
+}
+
+bool TConfigurePartsAtTable::HandleReply(TEvDataShard::TEvProposeTransactionResult::TPtr& ev, TOperationContext& context) {
+    LOG_INFO_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+                DebugHint() << " HandleReply " << ev->Get()->ToString()
+                            << ", at schemeshard: " << context.SS->SelfTabletId());
+
+    if (!NTableState::CollectProposeTransactionResults(OperationId, ev, context)) {
+        return false;
+    }
+
+    return true;
+}
+
+
+// NCdcStreamState::TProposeAtTable
+//
+TProposeAtTable::TProposeAtTable(TOperationId id)
+    : OperationId(id)
+{
+    IgnoreMessages(DebugHint(), {TEvDataShard::TEvProposeTransactionResult::EventType});
+}
+
+bool TProposeAtTable::ProgressState(TOperationContext& context) {
+    LOG_INFO_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+                DebugHint() << " ProgressState"
+                            << ", at schemeshard: " << context.SS->SelfTabletId());
+
+    const auto* txState = context.SS->FindTx(OperationId);
+    Y_ABORT_UNLESS(txState);
+    Y_ABORT_UNLESS(IsExpectedTxType(txState->TxType));
+
+    TSet<TTabletId> shardSet;
+    for (const auto& shard : txState->Shards) {
+        Y_ABORT_UNLESS(context.SS->ShardInfos.contains(shard.Idx));
+        shardSet.insert(context.SS->ShardInfos.at(shard.Idx).TabletID);
+    }
+
+    context.OnComplete.ProposeToCoordinator(OperationId, txState->TargetPathId, txState->MinStep, shardSet);
+    return false;
+}
+
+bool TProposeAtTable::HandleReply(TEvPrivate::TEvOperationPlan::TPtr& ev, TOperationContext& context) {
+    LOG_INFO_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+                DebugHint() << " HandleReply TEvOperationPlan"
+                            << ", step: " << ev->Get()->StepId
+                            << ", at schemeshard: " << context.SS->SelfTabletId());
+
+    const auto* txState = context.SS->FindTx(OperationId);
+    Y_ABORT_UNLESS(txState);
+    Y_ABORT_UNLESS(IsExpectedTxType(txState->TxType));
+    const auto& pathId = txState->TargetPathId;
+
+    Y_ABORT_UNLESS(context.SS->PathsById.contains(pathId));
+    auto path = context.SS->PathsById.at(pathId);
+
+    Y_ABORT_UNLESS(context.SS->Tables.contains(pathId));
+    auto table = context.SS->Tables.at(pathId);
+
+    table->AlterVersion += 1;
+
+    NIceDb::TNiceDb db(context.GetDB());
+    context.SS->PersistTableAlterVersion(db, pathId, table);
+
+    context.SS->ClearDescribePathCaches(path);
+    context.OnComplete.PublishToSchemeBoard(OperationId, pathId);
+
+    context.SS->ChangeTxState(db, OperationId, TTxState::ProposedWaitParts);
+    return true;
+}
+
+bool TProposeAtTable::HandleReply(TEvDataShard::TEvSchemaChanged::TPtr& ev, TOperationContext& context) {
+    LOG_INFO_S(context.Ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
+                DebugHint() << " TEvDataShard::TEvSchemaChanged"
+                            << " triggers early, save it"
+                            << ", at schemeshard: " << context.SS->SelfTabletId());
+
+    NTableState::CollectSchemaChanged(OperationId, ev, context);
+    return false;
+}
+
+
+// NCdcStreamState::TProposeAtTableDropSnapshot
+//
+bool TProposeAtTableDropSnapshot::HandleReply(TEvPrivate::TEvOperationPlan::TPtr& ev, TOperationContext& context) {
+    TProposeAtTable::HandleReply(ev, context);
+
+    const auto* txState = context.SS->FindTx(OperationId);
+    Y_ABORT_UNLESS(txState);
+    const auto& pathId = txState->TargetPathId;
+
+    Y_ABORT_UNLESS(context.SS->TablesWithSnapshots.contains(pathId));
+    const auto snapshotTxId = context.SS->TablesWithSnapshots.at(pathId);
+
+    auto it = context.SS->SnapshotTables.find(snapshotTxId);
+    if (it != context.SS->SnapshotTables.end()) {
+        it->second.erase(pathId);
+        if (it->second.empty()) {
+            context.SS->SnapshotTables.erase(it);
+        }
+    }
+
+    context.SS->SnapshotsStepIds.erase(snapshotTxId);
+    context.SS->TablesWithSnapshots.erase(pathId);
+
+    NIceDb::TNiceDb db(context.GetDB());
+    context.SS->PersistDropSnapshot(db, snapshotTxId, pathId);
+
+    context.SS->TabletCounters->Simple()[COUNTER_SNAPSHOTS_COUNT].Sub(1);
+    return true;
+}
+
+}  // NKikimr::NSchemeShard::NCdcStreamState
